Change the chain buffer to 1032 bits
The reasoning is the following: even though it was possible to program a
Spartan 6 class device with 1024 bits, if the chain includes another
device, e.g. an EEPROM, at lease one other bit will be needed to put
the device in BYPASS.

Eight bits is the minimum that we can add (one byte).